[WGSL] GlobalVariableRewriter should visit callees before entrypoint https://bugs.webkit.org/show_bug.cgi?id=257561 rdar://110077039
Reviewed by Myles C. Maxfield.

When collecting used globals, a function should collect its callees' used variables before
its own, but currently we have it inverted. As it currently stands, not only we might miss
globals used by callees, but we'd also add variables used by the caller to the callee. Fixing
it is just a matter of reordering the operations in the visitor.

* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::visit):
* Source/WebGPU/WGSL/tests/valid/global-used-by-callee.wgsl: Added.

Canonical link: https://commits.webkit.org/264791@main