[FreeType] Use builtin WOFF2 support https://bugs.webkit.org/show_bug.cgi?id=213862
Reviewed by Michael Catanzaro.

FreeType v2.10.2 added native support for WOFF2. Add a CMake check to
verify that the support was compiled in. Enable this support for the
PlayStation port and make sure to load the brotli module if needed.

The only function used from `WOFFFileFormat.h` is
`convertWOFFToSfntIfNecessary` so only expose that. Clean up the
implementation of `WOFFFileFormat.cpp` to reflect that.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebCore/platform/graphics/WOFFFileFormat.cpp:
(WebCore::convertWOFFToSfntIfNecessary):
* Source/WebCore/platform/graphics/WOFFFileFormat.h:
* Source/WebCore/platform/graphics/freetype/FontCustomPlatformDataFreeType.cpp:
(WebCore::FontCustomPlatformData::supportsFormat):
* Source/WebKit/WebProcess/EntryPoint/playstation/WebProcessMain.cpp:
(main):
* Source/cmake/OptionsPlayStation.cmake:
* Source/cmake/WOFF2Checks.cmake: Added.

Canonical link: https://commits.webkit.org/264815@main