Viewport layout size is sometimes incorrect when a site specifies an initial scale https://bugs.webkit.org/show_bug.cgi?id=257583 rdar://109483331
Reviewed by Dean Jackson.

On iOS, if a page specifies a viewport meta tag including an initial scale value that is < 1,
and if the `minimumEffectiveDeviceWidth` is non-zero, the resulting layout width will have an
undesirable value. Instead, the initial scale should be ignored when the `minimumEffectiveDeviceWidth`
is specified when computing the layout width.

This PR fixes this by ignoring the initial scale only when the `minimumEffectiveDeviceWidth` actually
has an effect on the viewport layout width.

Also refactors the `applyViewportArgument` to be simpler.

* Source/WebCore/page/ViewportConfiguration.cpp:
(WebCore::applyViewportArgument):
(WebCore::ViewportConfiguration::updateConfiguration):
(WebCore::ViewportConfiguration::layoutWidth const):
* Source/WebCore/page/ViewportConfiguration.h:
(WebCore::ViewportConfiguration::Parameters::operator== const):
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/ios/Viewport.mm: Added.
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/264842@main