Restore iOS sandbox profile generation in engineering builds https://bugs.webkit.org/show_bug.cgi?id=257276 rdar://109790387
Reviewed by Alexey Proskuryakov.

264752@main (6f04c95bf8f3) stopped preprocessing iOS sandbox profiles
unconditionally during at-desk builds. The logic was that the profiles
are only used during WebKit's production build, so there was no need to
generate them locally. However, it's helpful for engineers to be able to
manually test the sandbox profiles outside of a production build
environment.

When building for iOS, make WebKit depend on the "Sandbox Profiles"
target. Change that target to install sandbox profiles to
BUILT_PRODUCTS_DIR in non-install builds. Add a dependency on WTF, since
Platform headers are used to preprocess the sandbox profiles and the
target needs to be scheduled after they are available.

The result is that sandbox profiles are preprocessed at their old
DerivedSources location, and are copied to a path like

    WebKitBuild/Debug-iphoneos/usr/local/share/sandbox/...

Also, remove the old "Copy iOS Sandbox Profiles for Manual Sandboxing"
script phase, which was an unused workflow.

* Source/WebKit/Configurations/SandboxProfiles.xcconfig: Declare a path
  variable used to install profiles into BUILT_PRODUCTS_DIR in
  non-install builds. Add $(BUILT_PRODUCTS_DIR)/usr/local/include to the
  search path, to pick up WTF headers during engineering builds.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/264904@main