Reduce sizeof(ThreadSafeWeakPtr) https://bugs.webkit.org/show_bug.cgi?id=257779 rdar://110368849
Reviewed by NOBODY (OOPS!).

In 264919@main I made ThreadSafeWeakPtr keep two pointers to support multiple inheritance.
We need a pointer to the control block, but also a pointer to the object of the correct type.
This is needed because if you have A inherit from B, then casting an A* to a B* gives you a
different location in memory.  With multiple inheritance, the pointer to the object needs to
be stored in the smart pointer.

However, if we assume that our structures are less than 65kb, then the pointer can be reconstructed
from a base pointer which we already store in the control block and an offset that only requires
16 bits.  If you assume that the virtual address space will never be completely used (which we already
do for users of CompactRefPtrTuple) then you can store that offset in the same register as the pointer.

* Source/WTF/wtf/CompactRefPtrTuple.h:
* Source/WTF/wtf/ThreadSafeWeakHashSet.h:
* Source/WTF/wtf/ThreadSafeWeakPtr.h:
(WTF::ThreadSafeWeakPtrControlBlock::makeStrongReferenceIfPossible const):
(WTF::ThreadSafeWeakPtrControlBlock::objectOffset):
(WTF::ThreadSafeWeakPtr::ThreadSafeWeakPtr):
(WTF::ThreadSafeWeakPtr::operator=):
(WTF::ThreadSafeWeakPtr::get const):
(WTF::ThreadSafeWeakPtr::controlBlockAndObjectOffset):
(WTF::ThreadSafeWeakPtr::controlBlock): Deleted.
* Tools/TestWebKitAPI/Tests/WTF/WeakPtr.cpp:
(TestWebKitAPI::TEST):