[WGSL] Make callee ordering deterministic https://bugs.webkit.org/show_bug.cgi?id=257714 rdar://110263964
Reviewed by Myles C. Maxfield.

The call graph's callee list was built from a HashMap, which meant we visited
callees in a non-deterministic order. This makes it harder to write tests that
rely on the order of the output, and we also had crashes that were happening
only when we visited the callees in a certain order. To fix this, we consistently
put callees in a list as we discover them and use the previous hash map to keep
track of where a given callee is on the list so we can add call sites to it later.

* Source/WebGPU/WGSL/CallGraph.cpp:
(WGSL::CallGraphBuilder::visit):

Canonical link: https://commits.webkit.org/264932@main