[FFC] ASSERTION FAILED: formattingContextRoot.hasChild() when flex item has inline content https://bugs.webkit.org/show_bug.cgi?id=257802
Reviewed by Antti Koivisto.

With the (relatively) recent change of sharing Layout::Boxes across integration type of formatting contexts (e.g. for an inline-block box which is both a root of an IFC and a
leaf box in the parent IFC we only construct a single Layout::Box), we have to keep the integration object alive in order to not get the flex subtree destroyed including the flex items
which act as root boxes for the associated IFCs.

* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::layoutUsingFlexFormattingContext):
* Source/WebCore/rendering/RenderFlexibleBox.h:

Canonical link: https://commits.webkit.org/264944@main