Remove iOS-only NSUserDefaults-backed UA override https://bugs.webkit.org/show_bug.cgi?id=217927 rdar://70699433
Reviewed by Tim Horton.

This commit relegates the mechanism that reads an override UA from the
NSUserDefault [com.apple.WebFoundation UserAgent] behind a
linked-on-or-after check for its namesake SDKAlignedBehavior.

We do this because the NSUserDefaults-backed UA string is incompatible
with the modern need to compose the UA. Instead, we strongly prefer that
clients use existing API to set the application name (or UA).

* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.cpp:
(WTF::computeSDKAlignedBehaviors):
* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.h:
* Source/WebCore/platform/ios/UserAgentIOS.mm:
(WebCore::standardUserAgentWithApplicationName):

Canonical link: https://commits.webkit.org/265034@main