Modify JPEGXL detection for Windows https://bugs.webkit.org/show_bug.cgi?id=257898
Reviewed by Michael Catanzaro.

In 265024@main `USE_JPEGXL` was moved into `WebKitFeatures.cmake` and
exposed to all ports. Previously Windows would just detect libjxl and if
present turn on support. The libjxl library is only distributed on x64
so disable it for x86. Move the check for the library below the options
list and use `REQUIRED` to error if not found.

* Source/cmake/OptionsWin.cmake:

Canonical link: https://commits.webkit.org/265035@main