[Mac] Animation stutters when making a video full screen in Safari https://bugs.webkit.org/show_bug.cgi?id=257922 rdar://109325226
Reviewed by Simon Fraser.

In some circumstances, there may be a significant delay between when fullscreen is initiated
through -[WKFullscreenWindowController beganEnterFullScreenWithInitialFrame], and when the
animation to enter fullscreen begins
via -window:startCustomAnimationToEnterFullScreenWithDuration:. In this case, the intent was to put
in place the animation's transformations, but with a speed of 0 (so that the animation would not
progress). However, the speed parameter to the animation methods was ignored, and the animation
would use the default speed of 1. When the delay between the two functions mentioned above occurrs,
it has the effect of making it look like the animation starts early and then "restarts".

Actually obey the speed parameter in zoomAnimation() and maskAnimation().

* Source/WebKit/UIProcess/mac/WKFullScreenWindowController.mm:
(zoomAnimation):
(maskAnimation):

Canonical link: https://commits.webkit.org/265051@main