REGRESSION: YouTube resumes Playing during Paused Video when attempting to Select Text https://bugs.webkit.org/show_bug.cgi?id=257971 rdar://104320780
Reviewed by NOBODY (OOPS!).

In full-screen video, a `VCKImageAnalysisOverlayView` is added on top of the `WKWebView` in the
view hierarchy to facilitate Live Text interaction. Previously, the `VCKImageAnalysisOverlayView`
would handle any `mouseDown` events and not pass them on to the `WKWebView`. However, a change was
made such that the view now always calls `[super mouseDown:event]`, resulting in the `WKWebView`
errenously being given mouse down events.

As a result, a `click` JS event is now being dispatched when the Live Text selection is interacted with
because the mouse down event is given to WebKit.

This PR fixes this by early returning from the `mouseDown` and `mouseUp` methods if the interaction point
is within the `VCKImageAnalysisOverlayView` bounds, and if there is an interactable item underneath the
interaction point, or if there is an active text selection.

Also renamed the `imageAnalysisOverlayViewHasInteractableItemAtPoint` function to better reflect its
meaning.

Also added an API test to verify this correct behavior.

* Source/WebKit/UIProcess/mac/PageClientImplMac.mm:
(WebKit::PageClientImpl::setCursor):
* Source/WebKit/UIProcess/mac/WebViewImpl.h:
* Source/WebKit/UIProcess/mac/WebViewImpl.mm:
(WebKit::WebViewImpl::mouseDown):
(WebKit::WebViewImpl::mouseUp):
(WebKit::WebViewImpl::imageAnalysisOverlayViewShouldHandleMouseEvents const):
(WebKit::WebViewImpl::imageAnalysisOverlayViewHasInteractableItemAtPoint const):
(WebKit::WebViewImpl::imageAnalysisOverlayViewHasCursorAtPoint const): Deleted.
* Tools/TestWebKitAPI/Tests/WebKitCocoa/FullscreenVideoTextRecognition.mm:
(swizzleInteractableItemExistsAtPoint):
(-[FullscreenVideoTextRecognitionWebView initWithFrame:configuration:]):
(TestWebKitAPI::TEST):